backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
python-dotenv==0.19.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
tensorflow==2.6.0
statsmodels==0.13.2
python-multipart==0.0.5
mysql-connector-python==8.0.26
alembic==1.7.3
pydantic==1.8.2
requests==2.26.0
python-jose==3.3.0
passlib==1.7.4
h5py==3.1.0
joblib==1.0.1
pymysql==1.0.2
cryptography==3.4.8 